refactor(aio): move embedded components to EmbeddedModule
